Keep package builds and strict metric typing valid
Configure Hatch to ship the src package and make CAGR strict-type compatible so local verification can install and inspect the MVP core without packaging failures.\n\nConstraint: Task 1 scope is limited to pyproject.toml and domain/data/strategy/backtest/metrics core files.\nRejected: Add README or package initializer | broader file changes exceed the assigned Task 1 slice.\nConfidence: high\nScope-risk: narrow\nDirective: Keep MVP code offline-only and avoid broker/network/credential dependencies.\nTested: uv run package import smoke; ruff check scoped files; mypy --strict scoped files; compileall src; pytest attempted but no tests exist yet in current worktree.\nNot-tested: Full passing pytest suite awaits Task 3 test files.

+++ src/auto_trading_bot/metrics.py
@@ -33,7 +33,7 @@ def cagr(equity_curve: tuple[EquityPoint, ...], initial_cash: float) -> float | 
     if elapsed_days <= 0 or initial_cash <= 0:
         return None
     years = elapsed_days / 365.25
-    return (equity_curve[-1].equity / initial_cash) ** (1 / years) - 1.0
+    return float((equity_curve[-1].equity / initial_cash) ** (1 / years) - 1.0)
 
 
 def volatility(
